pyproject.toml
[project]
name = "voodradar"
version = "0.1.0"
description = "AI ad-intelligence platform for mobile-game publishers — Voodoo Hack 2026, Track 3."
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "polars>=1.10",
    "pyarrow>=18.0",
    "pillow>=11.0",
    "numpy>=2.1",
    "scikit-learn>=1.5",
    "weasyprint>=63.0",
    "tenacity>=9.0",
    "rich>=13.9",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29",
    "jupyter>=1.1",
    "jupyterlab>=4.3",
    "jupytext>=1.16",
    "ruff>=0.7",
    "pytest>=8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
